Upgrade to Ubuntu 17.10, and more
Env vars control CPU count, memory, and shared dir for VM.

Use Ubuntu kernel that ships with Ubuntu 17.10 (4.13), rather than
installing the latest mainline kernel. This gets rid of the need for the
vagrant-vbguest plugin, which could be annoying at times.

Use the version of tmux that is in the repositories (2.5) rather than
compiling it.

Use neovim from the official repositories rather than a PPA.